refactor(core): split utils.ts into paths/urls/dates/auth modules (#1116)
packages/core/src/core/utils.ts was 583 lines mixing four unrelated
topics: filesystem paths, URL parsing, date math, and auth. That made
it the project's largest junk drawer and the file most likely to grow
when someone needed "a small helper".

Split into focused modules, each independently testable:

- core/paths.ts (104 LOC) — getDataDir, getStatePath, getBackupDir,
  getCacheDir, getGistIdPath, getStateCachePath, stateFileExists,
  getCLIVersion. All ~/.oss-autopilot/ filesystem layout.
- core/urls.ts (98 LOC) — parseGitHubUrl, extractOwnerRepo, splitRepo,
  isOwnRepo + the ParsedGitHubUrl interface. https://github.com/
  prefix enforcement and owner/repo character validation live here.
- core/dates.ts (47 LOC) — daysBetween, formatRelativeTime,
  byDateDescending. No fs, no auth.
- core/auth.ts (139 LOC) — getGitHubToken, requireGitHubToken,
  resetGitHubTokenCache, getGitHubTokenAsync, detectGitHubUsername.
  The token cache (cachedGitHubToken / tokenFetchAttempted) and
  the GITHUB_USERNAME_RE pattern are now scoped to this module.
- core/concurrency.ts gains DEFAULT_CONCURRENCY and sleep alongside
  the existing runWorkerPool — they were the only non-topic-bound
  exports of the old utils.ts.

core/utils.ts is reduced to a 28-line back-compat shim that
re-exports everything from the new modules. Existing imports keep
working unchanged; new code should import from the focused module.
A follow-up issue tracks removing the shim in the next major.

No behavior change. All 2077 core + 256 dashboard + 181 mcp-server
tests pass without modification.
