+++ packages/core/src/core/concurrency.ts
@@ -1,4 +1,16 @@
 /**
+ * Concurrency primitives shared across modules.
+ */
+
+/** Default concurrency limit for parallel GitHub API requests. */
+export const DEFAULT_CONCURRENCY = 5;
+
+/** Async sleep — exported for mockability in tests. */
+export function sleep(ms: number): Promise<void> {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
+
+/**
  * Runs a worker pool that processes items with bounded concurrency.
  * N workers consume from a shared index. On any worker error, remaining
  * workers are aborted via a shared flag and the error is propagated.
